fix terrain texture palette hazard
When terrain textures are being updated they can release the last owner of the texture. This was a problem for the texture palette in the terrain editor as it earlier in the frame had asked for the indices of the texture SRVs so it could draw them.

This was causing GPU crashes (which WorldEdit would recover from thankfully). Now the renderer can be asked for any texture using a special function and it won't be queued for release until at least the next frame. Giving us a chance to safely use it in the current frame in ImGui.