change grid from depth bias to height bias
For whatever reason (quite possibly user error) the HW depth bias is now always producing far too much offset on my RTX 2080. WARP is fine and shows what I expect but that doesn't mean WARP is correct either.

If I had more GPUs to test on I'd investigate further and see which one is the odd one out. But barring access to such resources let's just approach the problem from a different perspective.

A 1mm offset to the grid height (oriented using the camera's height) produces enough difference that the grid doesn't Z-fight with terrain at the same height as it. Which is the whole reason the depth bias was there.

The best part is this requires no shader changes as we can just apply the bias to the height passed into the shader.